requirements.txt
numpy
opencv-python
matplotlib
black
flake8
isort
mkdocs
pip
python-dotenv
-e .
